output train set to jpg, refactoring
